+++ FirebaseFunctions/Backend/start.sh
@@ -55,6 +55,8 @@ FUNCTIONS_BIN="./node_modules/.bin/functions"
 "${FUNCTIONS_BIN}" deploy httpErrorTest --trigger-http
 "${FUNCTIONS_BIN}" deploy throwTest --trigger-http
 "${FUNCTIONS_BIN}" deploy timeoutTest --trigger-http
+"${FUNCTIONS_BIN}" deploy genStream --trigger-http
+"${FUNCTIONS_BIN}" deploy genStreamError --trigger-http
 
 if [ "$1" != "synchronous" ]; then
   # Wait for the user to tell us to stop the server.