+++ Firestore/core/include/firebase/firestore/timestamp.h
@@ -23,8 +23,8 @@
 #include <string>
 
 #if !defined(_STLPORT_VERSION)
-#include <chrono>  // NOLINT(build/c++11)
-#endif             // !defined(_STLPORT_VERSION)
+#include <chrono>
+#endif  // !defined(_STLPORT_VERSION)
 
 namespace firebase {