+++ Firestore/core/src/api/load_bundle_task.h
@@ -19,7 +19,7 @@
 #include <cstdint>
 #include <functional>
 #include <memory>
-#include <mutex>  // NOLINT(build/c++11)
+#include <mutex>
 #include <string>
 #include <utility>
 #include <vector>