+++ IntegrationTesting/ClientApp/Shared/objc-module-import-test.m
@@ -24,18 +24,19 @@
 #if TARGET_OS_IOS && !TARGET_OS_MACCATALYST
 @import FirebaseAppDistribution;
 #endif
+@import FirebaseAuth;
 @import FirebaseCore;
 @import FirebaseCrashlytics;
-@import FirebaseDatabase;
 #if TARGET_OS_IOS && !TARGET_OS_MACCATALYST
 @import FirebaseDynamicLinks;
+#endif
+#if (TARGET_OS_IOS && !TARGET_OS_MACCATALYST) || TARGET_OS_TV
+@import FirebasePerformance;
+@import FirebaseInAppMessaging;
 #endif
 @import FirebaseFirestore;
 @import FirebaseFunctions;
 @import FirebaseInstallations;
 @import FirebaseMessaging;
-#if (TARGET_OS_IOS && !TARGET_OS_MACCATALYST) || TARGET_OS_TV
-@import FirebasePerformance;
-@import FirebaseInAppMessaging;
-#endif
+@import FirebaseStorage;
 @import FirebaseRemoteConfig;